feat(vm): allow `none` `cdrom`
